requirements.txt
python-dotenv
langchain
langchain-groq
langgraph
python-telegram-bot